chore: decoupled filter at NewsTableViewController
